requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq>=0.9.0
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2
seaborn==0.13.2
python-multipart==0.0.12
pydantic==2.9.2
aiofiles==24.1.0
python-dotenv==1.0.1
faker==30.8.0
plotly>=5.24.0
scikit-learn>=1.5.0
